pipeline-manager: reorganize controller and transport types
