+++ crates/adapters/src/test/kafka.rs
@@ -6,7 +6,7 @@ use anyhow::{anyhow, bail, Result as AnyResult};
 use csv::WriterBuilder as CsvWriterBuilder;
 use futures::executor::block_on;
 use log::{error, info};
-use pipeline_types::transport::default_redpanda_server;
+use pipeline_types::transport::kafka::default_redpanda_server;
 use rdkafka::{
     admin::{AdminClient, AdminOptions, NewPartitions, NewTopic, TopicReplication},
     client::{Client, DefaultClientContext},
