+++ crates/pipeline-types/src/transport/http.rs
@@ -1,0 +1,27 @@
+use serde::Deserialize;
+use serde_json::Value as JsonValue;
+use utoipa::ToSchema;
+
+/// A set of updates to a SQL table or view.
+///
+/// The `sequence_number` field stores the offset of the chunk relative to the
+/// start of the stream and can be used to implement reliable delivery.
+/// The payload is stored in the `bin_data`, `text_data`, or `json_data` field
+/// depending on the data format used.
+#[derive(Deserialize, ToSchema)]
+pub struct Chunk {
+    pub sequence_number: u64,
+
+    // Exactly one of the following fields must be set.
+    // This should be an enum inlined with `#[serde(flatten)]`, but `utoipa`
+    // struggles to generate a schema for that.
+    /// Base64 encoded binary payload, e.g., bincode.
+    pub bin_data: Option<Vec<u8>>,
+
+    /// Text payload, e.g., CSV.
+    pub text_data: Option<String>,
+
+    /// JSON payload.
+    #[schema(value_type = Option<Object>)]
+    pub json_data: Option<JsonValue>,
+}
