+++ crates/pipeline-types/src/transport/mod.rs
@@ -1,0 +1,4 @@
+pub mod file;
+pub mod http;
+pub mod kafka;
+pub mod url;
