+++ crates/pipeline_manager/src/api.rs
@@ -175,13 +175,13 @@ request is rejected."
         pipeline_types::controller::config::ConnectorConfig,
         pipeline_types::controller::config::TransportConfig,
         pipeline_types::controller::config::FormatConfig,
-        pipeline_types::transport::FileInputConfig,
-        pipeline_types::transport::FileOutputConfig,
-        pipeline_types::transport::UrlInputConfig,
-        pipeline_types::transport::KafkaInputConfig,
-        pipeline_types::transport::KafkaOutputConfig,
-        pipeline_types::transport::KafkaLogLevel,
-        pipeline_types::transport::Chunk,
+        pipeline_types::transport::file::FileInputConfig,
+        pipeline_types::transport::file::FileOutputConfig,
+        pipeline_types::transport::url::UrlInputConfig,
+        pipeline_types::transport::kafka::KafkaInputConfig,
+        pipeline_types::transport::kafka::KafkaOutputConfig,
+        pipeline_types::transport::kafka::KafkaLogLevel,
+        pipeline_types::transport::http::Chunk,
         pipeline_types::csv::CsvEncoderConfig,
         pipeline_types::csv::CsvParserConfig,
         pipeline_types::json::JsonEncoderConfig,
